+++ WhiskyMan.Backend/WhiskyMan.Repositories.Interfaces/IBottleDescriptionRepository.cs
@@ -8,7 +8,7 @@ namespace WhiskyMan.Repositories.Interfaces
 {
     public interface IBottleDescriptionRepository
     {
-        Task<int> AddBottleDescription(BottleDescriptionForAddition bottleDescription);
+        Task<long> AddBottleDescription(BottleDescriptionForAddition bottleDescription);
         Task<List<BottleDescriptionReference>> GetActiveBottleDescriptionsReferences();
         Task<List<TagReference>> GetActiveTagReferences();
     }
