+++ WhiskyMan.Backend/WhiskyMan.Repositories.Interfaces/IBottleRepository.cs
@@ -9,7 +9,7 @@ namespace WhiskyMan.Repositories.Interfaces
 {
     public interface IBottleRepository
     {
-        Task<int> AddBottle(BottleForAddition bottle);
+        Task<long> AddBottle(BottleForAddition bottle);
         Task<List<BottleView>> GetActiveBottleViews();
     }
 }
